[ST7796S] Code base to programs the display
ST7796S initial configuration and send data to memory for display.

+++ kit-stm32f4/Standard_Library/STM32F4_ST7796-Display/configSPI.h
@@ -10,6 +10,6 @@ void Init_SPI(void);
 void SPI_writeReg(uint8_t reg, uint8_t regvalue);
 uint8_t SPI_readReg(uint8_t reg);
 void SPI_TxData(uint8_t value);
-void SPI_writeData(uint8_t reg, uint8_t *Datavalue, uint8_t length);
+void SPI_writeData(uint8_t reg, uint8_t *Datavalue, uint32_t length);
 
 #endif
